Move `NetworkService` traits to `sc-network`
