+++ client/cli/src/params/network_params.rs
@@ -18,8 +18,11 @@
 
 use crate::{arg_enums::SyncMode, params::node_key_params::NodeKeyParams};
 use clap::Args;
-use sc_network::{config::NetworkConfiguration, multiaddr::Protocol};
-use sc_network_common::config::{NodeKeyConfig, NonReservedPeerMode, SetConfig, TransportConfig};
+use sc_network::{
+	config::{NetworkConfiguration, SetConfig},
+	multiaddr::Protocol,
+};
+use sc_network_common::config::{NodeKeyConfig, NonReservedPeerMode, TransportConfig};
 use sc_service::{
 	config::{Multiaddr, MultiaddrWithPeerId},
 	ChainSpec, ChainType,
