+++ client/consensus/grandpa/src/communication/tests.rs
@@ -25,14 +25,14 @@ use super::{
 use crate::{communication::grandpa_protocol_name, environment::SharedVoterSetState};
 use futures::prelude::*;
 use parity_scale_codec::Encode;
-use sc_network::{config::Role, Multiaddr, PeerId, ReputationChange};
+use sc_network::{
+	config::Role, Multiaddr, NetworkBlock, NetworkEventStream, NetworkNotification, NetworkPeers,
+	NetworkSyncForkRequest, NotificationSenderError, NotificationSenderT as NotificationSender,
+	PeerId, ReputationChange,
+};
 use sc_network_common::{
 	config::MultiaddrWithPeerId,
 	protocol::{event::Event as NetworkEvent, role::ObservedRole, ProtocolName},
-	service::{
-		NetworkBlock, NetworkEventStream, NetworkNotification, NetworkPeers,
-		NetworkSyncForkRequest, NotificationSender, NotificationSenderError,
-	},
 	sync::{SyncEvent as SyncStreamEvent, SyncEventStream},
 };
 use sc_network_gossip::Validator;
