+++ client/consensus/grandpa/src/lib.rs
@@ -703,7 +703,7 @@ pub fn grandpa_peers_set_config(
 		// Notifications reach ~256kiB in size at the time of writing on Kusama and Polkadot.
 		max_notification_size: 1024 * 1024,
 		handshake: None,
-		set_config: sc_network_common::config::SetConfig {
+		set_config: sc_network::config::SetConfig {
 			in_peers: 0,
 			out_peers: 0,
 			reserved_nodes: Vec::new(),
