+++ client/informant/src/display.rs
@@ -20,12 +20,10 @@ use crate::OutputFormat;
 use ansi_term::Colour;
 use log::info;
 use sc_client_api::ClientInfo;
-use sc_network_common::{
-	service::NetworkStatus,
-	sync::{
-		warp::{WarpSyncPhase, WarpSyncProgress},
-		SyncState, SyncStatus,
-	},
+use sc_network::NetworkStatus;
+use sc_network_common::sync::{
+	warp::{WarpSyncPhase, WarpSyncProgress},
+	SyncState, SyncStatus,
 };
 use sp_runtime::traits::{Block as BlockT, CheckedDiv, NumberFor, Saturating, Zero};
 use std::{fmt, time::Instant};
