+++ client/informant/src/lib.rs
@@ -23,7 +23,8 @@ use futures::prelude::*;
 use futures_timer::Delay;
 use log::{debug, info, trace};
 use sc_client_api::{BlockchainEvents, UsageProvider};
-use sc_network_common::{service::NetworkStatusProvider, sync::SyncStatusProvider};
+use sc_network::NetworkStatusProvider;
+use sc_network_common::sync::SyncStatusProvider;
 use sp_blockchain::HeaderMetadata;
 use sp_runtime::traits::{Block as BlockT, Header};
 use std::{collections::VecDeque, fmt::Display, sync::Arc, time::Duration};
