+++ client/network-gossip/Cargo.toml
@@ -22,6 +22,7 @@ log = "0.4.17"
 lru = "0.8.1"
 tracing = "0.1.29"
 prometheus-endpoint = { package = "substrate-prometheus-endpoint", version = "0.10.0-dev", path = "../../utils/prometheus" }
+sc-network = { version = "0.10.0-dev", path = "../network/" }
 sc-network-common = { version = "0.10.0-dev", path = "../network/common" }
 sc-peerset = { version = "4.0.0-dev", path = "../peerset" }
 sp-runtime = { version = "7.0.0", path = "../../primitives/runtime" }
