+++ client/network-gossip/src/bridge.rs
@@ -340,14 +340,12 @@ mod tests {
 		future::poll_fn,
 	};
 	use quickcheck::{Arbitrary, Gen, QuickCheck};
+	use sc_network::{
+		NetworkBlock, NetworkEventStream, NetworkNotification, NetworkPeers,
+		NotificationSenderError, NotificationSenderT as NotificationSender,
+	};
 	use sc_network_common::{
-		config::MultiaddrWithPeerId,
-		protocol::role::ObservedRole,
-		service::{
-			NetworkBlock, NetworkEventStream, NetworkNotification, NetworkPeers,
-			NotificationSender, NotificationSenderError,
-		},
-		sync::SyncEventStream,
+		config::MultiaddrWithPeerId, protocol::role::ObservedRole, sync::SyncEventStream,
 	};
 	use sp_runtime::{
 		testing::H256,
