+++ client/network/src/config.rs
@@ -29,7 +29,7 @@ use prometheus_endpoint::Registry;
 pub use sc_network_common::{
 	config::{
 		MultiaddrWithPeerId, NodeKeyConfig, NonReservedPeerMode, NotificationHandshake, ProtocolId,
-		SetConfig, SyncMode, TransportConfig,
+		SyncMode, TransportConfig,
 	},
 	protocol::{role::Role, ProtocolName},
 	request_responses::{
@@ -41,6 +41,34 @@ pub use sc_network_common::{
 
 // `std` dependencies
 use std::{future::Future, iter, net::Ipv4Addr, path::PathBuf, pin::Pin, sync::Arc};
+
+/// Configuration for a set of nodes.
+#[derive(Clone, Debug)]
+pub struct SetConfig {
+	/// Maximum allowed number of incoming substreams related to this set.
+	pub in_peers: u32,
+
+	/// Number of outgoing substreams related to this set that we're trying to maintain.
+	pub out_peers: u32,
+
+	/// List of reserved node addresses.
+	pub reserved_nodes: Vec<MultiaddrWithPeerId>,
+
+	/// Whether nodes that aren't in [`SetConfig::reserved_nodes`] are accepted or automatically
+	/// refused.
+	pub non_reserved_mode: NonReservedPeerMode,
+}
+
+impl Default for SetConfig {
+	fn default() -> Self {
+		Self {
+			in_peers: 25,
+			out_peers: 75,
+			reserved_nodes: Vec::new(),
+			non_reserved_mode: NonReservedPeerMode::Accept,
+		}
+	}
+}
 
 /// Extension to [`SetConfig`] for sets that aren't the default set.
 ///
