+++ client/service/test/src/lib.rs
@@ -22,11 +22,10 @@ use futures::{task::Poll, Future, TryFutureExt as _};
 use log::{debug, info};
 use parking_lot::Mutex;
 use sc_client_api::{Backend, CallExecutor};
-use sc_network::{config::NetworkConfiguration, multiaddr};
-use sc_network_common::{
-	config::{MultiaddrWithPeerId, TransportConfig},
-	service::{NetworkBlock, NetworkPeers, NetworkStateInfo},
+use sc_network::{
+	config::NetworkConfiguration, multiaddr, NetworkBlock, NetworkPeers, NetworkStateInfo,
 };
+use sc_network_common::config::{MultiaddrWithPeerId, TransportConfig};
 use sc_network_sync::SyncingService;
 use sc_service::{
 	client::Client,
